# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: VoxMeet API
  description: |
    VoxMeet is a self-hosted communication suite — group calls, messaging, and screen sharing.
    This API covers authentication, room management, user profiles, and real-time signaling.

    For WebSocket signaling protocol, see [docs/signaling-protocol.md](./signaling-protocol.md).
  version: "0.1.0"
  contact:
    name: VoxMeet
    url: https://github.com/your-org/voxmeet

servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://voxmeet.example.com
    description: Production

paths:
  # ─── Health ───────────────────────────────────────────────────
  /api/v1/health:
    get:
      summary: Health check
      tags: [Health]
      operationId: health
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: VoxMeet API Gateway

  # ─── Auth ─────────────────────────────────────────────────────
  /api/v1/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  minLength: 1
                  maxLength: 64
                  example: alice
                email:
                  type: string
                  format: email
                  example: alice@example.com
                password:
                  type: string
                  minLength: 6
                  example: secure123
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"

  /api/v1/auth/login:
    post:
      summary: Authenticate and receive a JWT
      tags: [Auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Users ────────────────────────────────────────────────────
  /api/v1/me:
    get:
      summary: Get current user profile
      tags: [Users]
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Rooms ────────────────────────────────────────────────────
  /api/v1/rooms:
    get:
      summary: List all rooms
      tags: [Rooms]
      operationId: listRooms
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Create a new room
      tags: [Rooms]
      operationId: createRoom
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 128
                  example: Team Standup
                is_public:
                  type: boolean
                  default: false
                max_participants:
                  type: integer
                  default: 20
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/rooms/{id}:
    get:
      summary: Get room by ID
      tags: [Rooms]
      operationId: getRoom
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update room settings
      tags: [Rooms]
      operationId: updateRoom
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                is_public:
                  type: boolean
                max_participants:
                  type: integer
      responses:
        "200":
          description: Room updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete a room
      tags: [Rooms]
      operationId: deleteRoom
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Room deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── WebSocket ────────────────────────────────────────────────
  /api/v1/ws:
    get:
      summary: WebSocket signaling endpoint
      tags: [Signaling]
      description: |
        Upgrade to WebSocket for real-time signaling. Must pass JWT as query param.

        See [docs/signaling-protocol.md](./signaling-protocol.md) for message types.
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: JWT obtained from POST /api/v1/auth/login
      responses:
        "101":
          description: Switching Protocols — WebSocket connection established
        "401":
          description: Missing or invalid token

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
        user_id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
          nullable: true

    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        owner_id:
          type: string
          format: uuid
        is_public:
          type: boolean
        max_participants:
          type: integer
        created_at:
          type: string
          format: date-time

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string